BUG: KMap returns null when geting existed elements

+++ app/src/main/java/schedule_system/utils/AdjData.java
@@ -2,5 +2,5 @@ package schedule_system.utils;
 
 public record AdjData(
         String name,
-        double weight) {
+        int weight) {
 }
